% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_regions.R
\name{fetch_tpm}
\alias{fetch_tpm}
\title{Fetch the robust TPM values.}
\usage{
fetch_tpm(region, cell_line)
}
\arguments{
\item{region}{A \code{GRanges} object corresponding to the regions of
interest.}

\item{cell_line}{The cell line to extract the robust TSS TPM values.}
}
\value{
A \code{GRanges} corresponding to the subset of the \code{region}
        that overlaps at least on robust TSS with the mean robust TSS values
        for each range.
}
\description{
This function will first fetch the TPM values that overlap the regions
of interest from the robust TSS dataset (FantomTSS.hg19). When there is more
than one TSS for a genomic region, the mean value will be used. Only the
regions that overlaps at least one robust TSS will be returned.
}

