% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_regions.R
\name{get_all_tss_tpm}
\alias{get_all_tss_tpm}
\title{Get all the robust TSS TPM values}
\usage{
get_all_tss_tpm(cell_line = NULL, exclude = NULL)
}
\arguments{
\item{cell_line}{Limit the value returned to those of a single cell line. If
\code{NULL}, all cell lines are used. Default: \code{NULL}.}

\item{exclude}{Exclude a cell line. If \code{NULL}, all cell lines are used.
               Default: \code{NULL}}
}
\value{
A numeric vector of all the robut TSS from Fantom5
}
\description{
Get all the robust TSS TPM values
}

