% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_design.R
\name{get_encode_design}
\alias{get_encode_design}
\title{Produce the design for a specific target}
\usage{
get_encode_design(target, cell_line)
}
\arguments{
\item{target}{The name of the target.}
}
\value{
The design in \code{data.frame} format.
}
\description{
Produce the design for a specific target
}

