% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_design.R
\name{get_encode_targets}
\alias{get_encode_targets}
\title{Get the available targets for a cell_line}
\usage{
get_encode_targets(cell_line)
}
\arguments{
\item{cell_line}{The name of the cell line (biosample).}
}
\value{
A vector containing the names of the targets
}
\description{
This function will extract the names of all the targets that have at least
one bam file available on the ENCODE web site.
}

