% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_regions.R
\name{split_by_quantiles}
\alias{split_by_quantiles}
\title{Split a GRanges into quantiles.}
\usage{
split_by_quantiles(gr, quantiles, cell_line = NULL)
}
\arguments{
\item{gr}{The \code{GRanges} to extract quantile info from.}

\item{quantiles}{The tresholds for the quantiles.}

\item{cell_line}{The name of the column to use for splitting \code{gr}. If
                 \code{NULL}, the first metadata column is used. Default:
                 \code{NULL}.}
}
\value{
A \code{GRangesList} of the original \code{GRanges} splitted into
        quantiles.
}
\description{
Split a GRanges into quantiles.
}

